feat(single): 결과 출근시간 변경 이식 — recomputeWhatIf 공통화(부부/싱글 일관)
PR B에서 거래유형/예산 토글은 옮겼으나 출근시간 in-result 변경은 범위 외였음 → 부부와 완전 동일하게 이식.

- lib/diagnosis/whatif.ts(신규): recomputeWhatIf 공통화 — 부부 result-content 로컬 함수를
  verbatim 추출(동작 동일). 출발시각 변경 시 baseline 통근 '분' 재추정(routePath·순위·여가 보존).
- result-content.tsx: 로컬 recomputeWhatIf 제거 → whatif.ts import(미사용 haversine import 정리).
  부부 시간/통근 what-if 동작 무변경(동일 함수).
- single-result-view.tsx: 출근시간을 "표시"에서 "변경 가능" 칩으로. baselineRef 캡처 +
  handleTimeWhatIf(recomputeWhatIf, 싱글=직장 1개 coordB=null) + TimeChipOptions 인라인.
  거래유형/예산 토글(통근 캐시 재활용)과 출근시간(통근 재계산) 경로 구분 유지.

검증(tsx, 커밋 안 함): tsc 0 / lint 0 error.
- 싱글 출근시간 변경 → baseline 통근 재추정(30→34분), commuteB 없이 안전, maxCommute 필터 정상.
- 부부 recomputeWhatIf 사용 유지(2곳), haversine 미사용 import 제거.
- 거래유형/예산/출근시간 셋 다 결과에서 변경 가능 → 부부/싱글 완전 일관.

거래유형/예산 토글(PR B)·통근 캐시·refilterPool 무변경. 기본 출근시간(08:00) 미선택 동작 유지.
★ 디자인/네트워크는 토큰+로직 검증 — 머지 후 브라우저 확인 권장.
